+++ NEWS.md
@@ -1,7 +1,7 @@
 > Copyright © 2016 RTE Reseau de transport d’electricite
 
 
-# antaresEditObject 0.9.4.9000
+# antaresEditObject 1.0.0
 
 NEW FEATURES :